# FSA.Tickets-Please/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "SIMS incident/signals management API",
    "version": "v1"
  },
  "paths": {
...
    "/api/Incidents/{id}": {
      "get": {
        "tags": [
          "Incidents"
        ],
        "summary": "Get incident by id",
        "operationId": "GetIncident",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/IncidentsDisplayModel"
            }
          },
          "500": {
            "description": "Server Error"
          }
        }
      }
    },
    "/api/Incidents": {
      "put": {
        "tags": [
          "Incidents"
        ],
        "summary": "Replace an incident",
        "operationId": "UpdateIncident",
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/IncidentUpdateModel"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/BaseIncident"
            }
          },
          "500": {
            "description": "Server Error"
          }
        }
      },
      "post": {
        "tags": [
          "Incidents"
        ],
        "summary": "Create an incident",
        "operationId": "CreateIncident",
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/IncidentCreateModel"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/BaseIncident"
            }
          },
          "500": {
            "description": "Server Error"
          }
        }
      }
    },
    "/api/Incidents/Classification/{id}/{classificationId}": {
      "post": {
        "tags": [
          "Incidents"
        ],
        "summary": "Update classification of an incident",
        "operationId": "UpdateClassification",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "path",
            "name": "classificationId",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/BaseIncident"
            }
          },
          "500": {
            "description": "Server Error"
          }
        }
      }
    },
    "/api/Incidents/Status/{id}/{statusId}": {
      "post": {
        "tags": [
          "Incidents"
        ],
        "summary": "Update status of an incident",
        "operationId": "UpdateStatus",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "path",
            "name": "statusId",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/BaseIncident"
            }
          },
          "500": {
            "description": "Server Error"
          }
        }
      }
    },
    "/api/Incidents/CloseAll": {
      "post": {
        "tags": [
          "Incidents"
        ],
        "summary": "Close list of incidents",
        "operationId": "CloseAll",
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/IncidentCloseViewModel"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "500": {
            "description": "Server Error"
          }
        }
      }
    },
    "/api/Incidents/LeadOfficer": {
      "post": {
        "tags": [
          "Incidents"
        ],
        "summary": "Update incident(s) lead officer",
        "operationId": "UpdateLeadOfficer",
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/UpdateLeadOfficerModel"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "500": {
            "description": "Server Error"
          }
        }
      }
    },
    "/api/Incidents/Dashboard": {
      "post": {
        "tags": [
          "Incidents"
        ],
        "summary": "Incident dashboard search",
        "operationId": "GetIncidentDashboard",
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/DashboardSearchViewModel"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/IncidentDashboardItem"
            }
          },
          "500": {
            "description": "Server Error"
          }
        }
      }
    },
    "/api/Links/{incidentSignal}/{id}": {
      "get": {
        "tags": [
          "Links"
        ],
        "summary": "Get All links for incident/signal",
        "operationId": "GetLinks",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "incidentSignal",
            "required": true,
            "type": "string"
          },
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "403": {
            "description": "Forbidden",
            "schema": {
              "type": "string"
            }
          },
          "500": {
            "description": "Server Error"
          }
        }
      }
    },
    "/api/Links/{incidentSignal}": {
      "delete": {
        "tags": [
          "Links"
        ],
        "summary": "Remove links for incident/signal",
        "operationId": "RemoveLink",
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "incidentSignal",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/UnlinkModel"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "500": {
            "description": "Server Error"
          },
          "403": {
            "description": "Forbidden",
            "schema": {
              "type": "string"
            }
          }
        }
      },
      "post": {
        "tags": [
          "Links"
        ],
        "summary": "Link two or more incidents",
        "operationId": "AddLink",
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "incidentSignal",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/LinkModel"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "403": {
            "description": "Forbidden",
            "schema": {
              "type": "string"
            }
          },
          "500": {
            "description": "Server Error"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "IncidentsDisplayModel": {
      "type": "object",
      "properties": {
        "category": {
          "type": "string",
          "readOnly": true
        },
        "signalStatus": {
          "type": "string",
          "readOnly": true
        },
        "incidentStatus": {
          "type": "string",
          "readOnly": true
        },
        "notifier": {
          "type": "string",
          "readOnly": true
        },
        "principalFBO": {
          "type": "string",
          "readOnly": true
        },
        "fboEmail": {
          "type": "string",
          "readOnly": true
        },
        "fboPhone": {
          "type": "string",
          "readOnly": true
        },
        "fboAddressLine1": {
          "type": "string",
          "readOnly": true
        },
        "fboAddressLine2": {
          "type": "string",
          "readOnly": true
        },
        "fboAddressTown": {
          "type": "string",
          "readOnly": true
        },
        "fboAddressPostcode": {
          "type": "string",
          "readOnly": true
        },
        "priority": {
          "type": "string",
          "readOnly": true
        },
        "classification": {
          "type": "string",
          "readOnly": true
        },
        "dataSource": {
          "type": "string",
          "readOnly": true
        },
        "productType": {
          "type": "string",
          "readOnly": true
        },
        "leadLocalAuthority": {
          "type": "string",
          "readOnly": true
        },
        "deathIllness": {
          "type": "string",
          "readOnly": true
        },
        "adminLead": {
          "type": "string",
          "readOnly": true
        },
        "id": {
          "type": "string",
          "readOnly": true
        },
        "commonId": {
          "format": "int32",
          "type": "integer",
          "readOnly": true
        },
        "mostUniqueId": {
          "format": "uuid",
          "type": "string",
          "readOnly": true
        },
        "incidentTitle": {
          "type": "string",
          "readOnly": true
        },
        "incidentTypeId": {
          "format": "int32",
          "type": "integer",
          "readOnly": true
        },
        "contactMethodId": {
          "format": "int32",
          "type": "integer",
          "readOnly": true
        },
        "statusId": {
          "format": "int32",
          "type": "integer",
          "readOnly": true
        },
        "signalStatusId": {
          "format": "int32",
          "type": "integer",
          "readOnly": true
        },
        "notifierId": {
          "format": "int32",
          "type": "integer",
          "readOnly": true
        },
        "principalFBOId": {
          "format": "int32",
          "type": "integer",
          "readOnly": true
        },
        "priorityId": {
          "format": "int32",
          "type": "integer",
          "readOnly": true
        },
        "classificationId": {
          "format": "int32",
          "type": "integer",
          "readOnly": true
        },
        "dataSourceId": {
          "format": "int32",
          "type": "integer",
          "readOnly": true
        },
        "signalUrl": {
          "type": "string",
          "readOnly": true
        },
        "productTypeId": {
          "format": "int32",
          "type": "integer",
          "readOnly": true
        },
        "leadOfficer": {
          "type": "string",
          "readOnly": true
        },
        "leadOffice": {
          "type": "string",
          "readOnly": true
        },
        "adminLeadId": {
          "format": "int32",
          "type": "integer",
          "readOnly": true
        },
        "fieldOfficer": {
          "type": "string",
          "readOnly": true
        },
        "leadLocalAuthorityId": {
          "format": "int32",
          "type": "integer",
          "readOnly": true
        },
        "laAdvised": {
          "type": "boolean",
          "readOnly": true
        },
        "deathIllnessId": {
          "format": "int32",
          "type": "integer",
          "readOnly": true
        },
        "receivedOn": {
          "format": "date-time",
          "type": "string",
          "readOnly": true
        },
        "incidentCreated": {
          "format": "date-time",
          "type": "string",
          "readOnly": true
        },
        "incidentClosed": {
          "format": "date-time",
          "type": "string",
          "readOnly": true
        },
        "lastChangedBy": {
          "type": "string",
          "readOnly": true
        },
        "lastChangedDate": {
          "format": "date-time",
          "type": "string",
          "readOnly": true
        }
      }
    },
    "IncidentUpdateModel": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "incidentTitle": {
          "type": "string"
        },
        "incidentDescription": {
          "type": "string"
        },
        "incidentTypeId": {
          "format": "int32",
          "type": "integer"
        },
        "contactMethodId": {
          "format": "int32",
          "type": "integer"
        },
        "statusId": {
          "format": "int32",
          "type": "integer"
        },
        "priorityId": {
          "format": "int32",
          "type": "integer"
        },
        "signalStatusId": {
          "format": "int32",
          "type": "integer"
        },
        "notifierId": {
          "format": "int32",
          "type": "integer"
        },
        "principalFBOId": {
          "format": "int32",
          "type": "integer"
        },
        "leadLocalAuthorityId": {
          "format": "int32",
          "type": "integer"
        },
        "classificationId": {
          "format": "int32",
          "type": "integer"
        },
        "dataSourceId": {
          "format": "int32",
          "type": "integer"
        },
        "signalUrl": {
          "type": "string"
        },
        "productTypeId": {
          "format": "int32",
          "type": "integer"
        },
        "leadOfficer": {
          "type": "string"
        },
        "leadOffice": {
          "type": "string"
        },
        "adminLeadId": {
          "format": "int32",
          "type": "integer"
        },
        "fieldOfficer": {
          "type": "string"
        },
        "laAdvised": {
          "type": "boolean"
        },
        "deathIllnessId": {
          "format": "int32",
          "type": "integer"
        }
      }
    },
    "BaseIncident": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "readOnly": true
        },
        "commonId": {
          "format": "int32",
          "type": "integer",
          "readOnly": true
        },
        "mostUniqueId": {
          "format": "uuid",
          "type": "string",
          "readOnly": true
        },
        "incidentTitle": {
          "type": "string",
          "readOnly": true
        },
        "incidentTypeId": {
          "format": "int32",
          "type": "integer",
          "readOnly": true
        },
        "contactMethodId": {
          "format": "int32",
          "type": "integer",
          "readOnly": true
        },
        "statusId": {
          "format": "int32",
          "type": "integer",
          "readOnly": true
        },
        "signalStatusId": {
          "format": "int32",
          "type": "integer",
          "readOnly": true
        },
        "notifierId": {
          "format": "int32",
          "type": "integer",
          "readOnly": true
        },
        "principalFBOId": {
          "format": "int32",
          "type": "integer",
          "readOnly": true
        },
        "priorityId": {
          "format": "int32",
          "type": "integer",
          "readOnly": true
        },
        "classificationId": {
          "format": "int32",
          "type": "integer",
          "readOnly": true
        },
        "dataSourceId": {
          "format": "int32",
          "type": "integer",
          "readOnly": true
        },
        "signalUrl": {
          "type": "string",
          "readOnly": true
        },
        "productTypeId": {
          "format": "int32",
          "type": "integer",
          "readOnly": true
        },
        "leadOfficer": {
          "type": "string",
          "readOnly": true
        },
        "leadOffice": {
          "type": "string",
          "readOnly": true
        },
        "adminLeadId": {
          "format": "int32",
          "type": "integer",
          "readOnly": true
        },
        "fieldOfficer": {
          "type": "string",
          "readOnly": true
        },
        "leadLocalAuthorityId": {
          "format": "int32",
          "type": "integer",
          "readOnly": true
        },
        "laAdvised": {
          "type": "boolean",
          "readOnly": true
        },
        "deathIllnessId": {
          "format": "int32",
          "type": "integer",
          "readOnly": true
        },
        "receivedOn": {
          "format": "date-time",
          "type": "string",
          "readOnly": true
        },
        "incidentCreated": {
          "format": "date-time",
          "type": "string",
          "readOnly": true
        },
        "incidentClosed": {
          "format": "date-time",
          "type": "string",
          "readOnly": true
        },
        "lastChangedBy": {
          "type": "string",
          "readOnly": true
        },
        "lastChangedDate": {
          "format": "date-time",
          "type": "string",
          "readOnly": true
        }
      }
    },
    "IncidentCreateModel": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "incidentTitle": {
          "type": "string"
        },
        "incidentDescription": {
          "type": "string"
        },
        "incidentTypeId": {
          "format": "int32",
          "type": "integer"
        },
        "contactMethodId": {
          "format": "int32",
          "type": "integer"
        },
        "statusId": {
          "format": "int32",
          "type": "integer"
        },
        "signalStatusId": {
          "format": "int32",
          "type": "integer"
        },
        "notifierId": {
          "format": "int32",
          "type": "integer"
        },
        "principalFBOId": {
          "format": "int32",
          "type": "integer"
        },
        "classificationId": {
          "format": "int32",
          "type": "integer"
        },
        "priorityId": {
          "format": "int32",
          "type": "integer"
        },
        "dataSourceId": {
          "format": "int32",
          "type": "integer"
        },
        "signalUrl": {
          "type": "string"
        },
        "productTypeId": {
          "format": "int32",
          "type": "integer"
        },
        "leadOfficer": {
          "type": "string"
        },
        "leadOffice": {
          "type": "string"
        },
        "adminLeadId": {
          "format": "int32",
          "type": "integer"
        },
        "fieldOfficer": {
          "type": "string"
        },
        "leadLocalAuthorityId": {
          "format": "int32",
          "type": "integer"
        },
        "laAdvised": {
          "type": "boolean"
        },
        "deathIllnessId": {
          "format": "int32",
          "type": "integer"
        },
        "receivedOn": {
          "format": "date-time",
          "type": "string"
        },
        "incidentCreated": {
          "format": "date-time",
          "type": "string"
        }
      }
    },
    "IncidentCloseViewModel": {
      "type": "object",
      "properties": {
        "incidentIds": {
          "type": "array",
          "items": {
            "format": "int32",
            "type": "integer"
          }
        }
      }
    },
    "UpdateLeadOfficerModel": {
      "required": [
        "ids",
        "officer"
      ],
      "type": "object",
      "properties": {
        "ids": {
          "type": "array",
          "items": {
            "format": "int32",
            "type": "integer"
          }
        },
        "officer": {
          "type": "string"
        }
      }
    },
    "DashboardSearchViewModel": {
      "type": "object",
      "properties": {
        "search": {
          "type": "string"
        },
        "pageSize": {
          "format": "int32",
          "type": "integer"
        },
        "pageNo": {
          "format": "int32",
          "type": "integer"
        }
      }
    },
    "IncidentDashboardItem": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "readOnly": true
        },
        "commonId": {
          "format": "int32",
          "type": "integer"
        },
        "priority": {
          "type": "string"
        },
        "title": {
          "type": "string"
        },
        "notifier": {
          "type": "string"
        },
        "leadOfficer": {
          "type": "string"
        },
        "status": {
          "type": "string"
        },
        "updated": {
          "format": "date-time",
          "type": "string"
        },
        "links": {
          "type": "array",
          "items": {
            "format": "int32",
            "type": "integer"
          }
        }
      }
    },
    "UnlinkModel": {
      "required": [
        "fromId",
        "toId"
      ],
      "type": "object",
      "properties": {
        "fromId": {
          "format": "int32",
          "type": "integer"
        },
        "toId": {
          "format": "int32",
          "type": "integer"
        }
      }
    },
    "LinkModel": {
      "required": [
        "fromId",
        "toIds"
      ],
      "type": "object",
      "properties": {
        "fromId": {
          "format": "int32",
          "type": "integer"
        },
        "toIds": {
          "type": "array",
          "items": {
            "format": "int32",
            "type": "integer"
          }
        },
        "comment": {
          "type": "string"
        }
      }
    },
...
  }
}